refactor: removed and refactored files added by vue-cli and refactored folder structure
