+++ web/src/App.vue
@@ -1,8 +1,7 @@
 <template>
   <div id="app">
     <div id="nav">
-      <router-link to="/">Home</router-link> |
-      <router-link to="/about">About</router-link>
+      <router-link to="/">Index</router-link>
     </div>
     <router-view/>
   </div>
