+++ web/src/main.js
@@ -1,7 +1,7 @@
 import Vue from "vue";
 import App from "./App.vue";
-import router from "./router";
-import store from "./store";
+import router from "./router/index";
+import store from "./store/index";
 
 Vue.config.productionTip = false;
