+++ web/src/router/index.js
@@ -1,0 +1,16 @@
+import Vue from "vue";
+import Router from "vue-router";
+import Index from "../views/Index.vue";
+
+Vue.use(Router);
+
+export default new Router({
+  routes: [
+    {
+      path: "/",
+      name: "index",
+      component: Index
+    }
+  ],
+  mode: "history"
+});
